requirements.txt
beautifulsoup4==4.12.2
discord.py==2.3.2
requests==2.31.0
